--- src/components/Footer.js
@@ -1,14 +1,0 @@
-import React  from 'react';
-import Logo from './logo2.png';
-
-import './Footer.scss';
-
-const Footer = () => {
-    return(
-      <div className="FooterClass">
-        <p className="FooterText">&copy; created by <a href= "https://akai.org.pl/" target="_blank" className="FooterText"> AKAI</a></p>
-      </div>
-    )
-  
-  }
-  export default Footer;
