+++ src/components/Footer/index.js
@@ -1,0 +1,15 @@
+import React  from 'react';
+
+import './index.scss';
+
+const Footer = () => {
+  return(
+    <div className="FooterClass">
+      <p className="FooterText">&copy; created by 
+        <a href= "https://akai.org.pl/" target="_blank" className="FooterText"> AKAI</a>
+      </p>
+    </div>
+  )
+}
+
+export default Footer;
