+++ src/components/LectureComponent/index.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import './LectureComponent.scss';
-import StarsComponent from './StarsComponent';
 import StarRatingComponent from "react-star-rating-component";
 
-import Person from "./Person";
+import './index.scss';
 
-class LectureComponent extends React.Component {
+import Person from "../Person";
+
+class LectureComponent extends Component {
 
   constructor(props) {
     super(props);
