+++ src/components/LectureComponent/index.scss
