+++ src/components/Person/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
-import './Person.scss';
+
+import './index.scss';
 
 const Person = ({ name, surname, profession }) => {
   return (
     <div className="lecture">
       <p className="lecture__profession">{profession}</p>
-      <p className="lecture__name">{name} {surname}</p>
+      <p className="lecture__name">{`${name} ${surname}`}</p>
     </div>
   )
 }
