+++ src/components/Person/index.scss
