+++ src/components/StarsComponent/index.scss
