+++ src/components/Tail/index.scss
