+++ src/containers/LectureContainer/index.scss
@@ -6,4 +6,5 @@
     color: rgb(255, 255, 255);
     font-size: 25px;
     margin-bottom: 100px;
+    width: 60vw;
 }
