+++ src/pages/LecturePage/index.scss
