+++ src/pages/MainPage/index.scss
@@ -1,0 +1,14 @@
+div, span
+{
+    margin: 0;
+    padding: 0;
+}
+
+.MainPage-Header
+{
+    width: 100vw;
+    background:#121212;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
